Modify class names and comments to improve readability
The command class names use a short form (Cmd for command).
Additionally, comments are not helpful since they describe what the
program is doing.

Not everyone will understand what Cmd actually means, since it is not
a short form that is widely recognized by everyone. Comments should
not describe a program, but rather, aid other readers of the code to
understand the program better.

The command class names will change Cmd -> Command, so that the full
name is spelt out. Comments which are redundant and descriptive are
removed, or rephrased to help the reader understand the program.

This will improve overall readability of the program, which is one of
the main goals of programming both in a team and personal environment.
