+++ src/main/java/drivers/TaskList.java
@@ -59,7 +59,7 @@ public class TaskList {
      */
     public Task markTask(int oneIndexedNum) throws TumuException {
         if (oneIndexedNum < 1 || oneIndexedNum > userTasks.size()) {
-            //Specified index from user is out of bounds of list.
+            // Specified index from user is out of bounds of list.
             if (userTasks.isEmpty()) {
                 throw new NoTaskException();
             } else {
@@ -80,7 +80,7 @@ public class TaskList {
      */
     public Task unmarkTask(int oneIndexedNum) throws TumuException {
         if (oneIndexedNum < 1 || oneIndexedNum > userTasks.size()) {
-            //Specified index from user is out of bounds of list.
+            // Specified index from user is out of bounds of list.
             if (userTasks.isEmpty()) {
                 throw new NoTaskException();
             } else {
@@ -101,7 +101,7 @@ public class TaskList {
      */
     public Task deleteTask(int oneIndexedNum) throws TumuException {
         if (oneIndexedNum < 1 || oneIndexedNum > userTasks.size()) {
-            //Specified index from user is out of bounds of list.
+            // Specified index from user is out of bounds of list.
             if (userTasks.isEmpty()) {
                 throw new NoTaskException();
             } else {
