+++ src/main/java/drivers/Tumu.java
@@ -77,7 +77,6 @@ public class Tumu extends Application {
      * @return a label with the specified text that has word wrap enabled.
      */
     private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
         Label textToAdd = new Label(text);
         textToAdd.setWrapText(true);
 
@@ -96,9 +95,9 @@ public class Tumu extends Application {
 
     @Override
     public void start(Stage stage) {
-        //Step 1. Setting up required components
+        // Step 1. Setting up required components
 
-        //The container for the content of the chat to scroll.
+        // The container for the content of the chat to scroll.
         scrollPane = new ScrollPane();
         dialogContainer = new VBox();
         scrollPane.setContent(dialogContainer); //attach scroll to dialog.
@@ -109,7 +108,7 @@ public class Tumu extends Application {
         AnchorPane mainLayout = new AnchorPane();
         mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
 
-        //Step 2. Formatting the window to look as expected
+        // Step 2. Formatting the window to look as expected
         stage.setTitle("Tumu");
         stage.setResizable(false);
         stage.setMinHeight(600.0);
@@ -139,7 +138,7 @@ public class Tumu extends Application {
         AnchorPane.setLeftAnchor(userInput , 1.0);
         AnchorPane.setBottomAnchor(userInput, 1.0);
 
-        //Step 3. Add functionality to handle user input.
+        // Step 3. Add functionality to handle user input.
         sendButton.setOnMouseClicked((event) -> {
             handleUserInput();
         });
@@ -147,7 +146,7 @@ public class Tumu extends Application {
         userInput.setOnAction((event) -> {
             handleUserInput();
         });
-        //Scroll down to the end every time dialogContainer's height changes.
+        // Scroll down to the end every time dialogContainer's height changes.
         dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
 
         scene = new Scene(mainLayout);
