+++ src/main/java/exceptions/GeneralException.java
@@ -4,6 +4,7 @@ package exceptions;
  * General exception to print an exception message directly.
  */
 public class GeneralException extends TumuException {
+
     /**
      * Constructor for the GeneralException class.
      * @param message Message to be displayed to the user.
